requirements.txt
torch>=2.1
torchvision>=0.16
transformers>=4.41
coremltools>=7.0
numpy>=1.23
Pillow>=10.0
